net: fec: hard phy reset on open
Some phys (e.g. LAN8720) need a stable clock while
and after a hard reset.

Before this patch, that was not the case as the driver disabled the
clock in probe() and enabled the clock again in open().

A solution that prevents disabling the clocks was considered,
but discarded because of bad power saving behavior.

This patch saves the reset dt properties on probe and does a reset on
every open() after clocks are enabled, to make sure the clock is stable
while and after a hard reset.

Reworked from Manfred Schlaegl's patch:
https://patchwork.kernel.org/patch/9391783/

Signed-off-by: Peter Fink <pfink@christ-es.de>