net: fec: Re-add a phy reset in fec_get_reset_phy
This commit fixes a bug introduced in 'net: fec: hard phy reset on open'
[6630625766c7d81d89a8b8479a3130d7f1ded7d9].

In configurations where there is no phy-handle configured in the devicetree
the network was unusable because no phy was detected.

This re-adds the reset in fec_get_reset_phy which was removed in the
previous commit.

Signed-off-by: Peter Fink <pfink@christ-es.de>
Signed-off-by: Alexander Schartner <aschartner@christ-es.de>