rtc: pcf8563: Prevent system lockups due to wrong data in rtc
The system might lockup with "systemd[1]: Time has been changed".
This can happen when a time > 2038-01-19 03:14:07 is read from the
rtc after the battery has died or at production time.

The battery low flag is already checked, but the time was still
returned to the system in any case.

Return -EINVAL in this case, although timedatectl and hwclock -r
will return error messages in this case until the time is set
manually or through ntp again.

This does not prevent the Y2K38 bug, but increases system stability
in case of an empty battery.

Signed-off-by: Peter Fink <pfink@christ-es.de>