video: backlight: pwm-backlight: add display enable gpio
This commit adds a possibility to supply a display enable
gpio through dtb to improve the visible display handling.

Previously there was a small gap where the backlight
was switched on simultaneously with the display logic
causing the display to float in an undefined state
until the display shows the correct image.

With this patch the display enable is activated and after
a delay of 150ms the backlight enable is switched on.
This results in the floating state being less visible or
not visible at all.

Signed-off-by: Peter Fink <pfink@christ-es.de>