ARM: dts: crix-arx: add own VDDD-supply regulator for sgtl5000
sgtl5000 failed to initialize correctly, because a LDO voltage
of 3.3V assigned to VDDD-supply causes an abort of the driver init
probe. (Limit for VDDD is 1.9V)

If no regulator is supplied for VDDD-supply the internal regulator
is used. This should be prevented as stated in ER1 erratum of sgtl500.

As crix-arx uses a LDO with 1.3V add a fixed regulator to the
devicetree to correctly initialize sgtl5000 with external LDO.

Signed-off-by: Steffen Kothe <skothe@christ-es.de>
Reviewed-by: Peter Fink <pfink@christ-es.de>