ARM: dts: blix-bx: fix property names for rotary-switch
The property names for the rotary switch in the 4.17.x kernel
driver is different to the one of the 4.1.x driver.
Yet the .dtsi file has been copied over without changes.
Thus the driver did not find the property and used default
values for the key created key events.

This commit syncs the driver to the .dtsi file and makes
the driver read the intended values again.